--- ESP32/DIY-Flow-Bench/tools/mafDataBosch067.json
@@ -1,258 +1,0 @@
-{
-    "0" : 0, 
-    "20" : 73.6, 
-    "40" : 71.6, 
-    "60" : 69.6, 
-    "80" : 67.6, 
-    "100" : 65.6, 
-    "120" : 62.6, 
-    "140" : 61.6, 
-    "160" : 59.6, 
-    "180" : 57.6, 
-    "200" : 55.6, 
-    "210" : 53.6, 
-    "230" : 51.6, 
-    "250" : 50.6, 
-    "270" : 48.6, 
-    "290" : 47.6, 
-    "310" : 45.6, 
-    "330" : 43.6, 
-    "350" : 42.6, 
-    "370" : 40.6, 
-    "390" : 39.6, 
-    "410" : 37.6, 
-    "430" : 36.6, 
-    "450" : 34.6, 
-    "470" : 33.6, 
-    "490" : 31.6, 
-    "510" : 30.6, 
-    "530" : 29.6, 
-    "550" : 27.6, 
-    "570" : 26.6, 
-    "590" : 25.1, 
-    "610" : 24.1, 
-    "630" : 22.9, 
-    "640" : 21.6, 
-    "660" : 20.6, 
-    "680" : 19.1, 
-    "700" : 18.1, 
-    "720" : 16.9, 
-    "740" : 15.8, 
-    "760" : 14.7, 
-    "780" : 13.6, 
-    "800" : 12.6, 
-    "820" : 11.6, 
-    "840" : 10.5, 
-    "860" : 9.4, 
-    "880" : 8.3, 
-    "900" : 7.3, 
-    "920" : 6.2, 
-    "940" : 4.8, 
-    "960" : 4.0, 
-    "980" : 1.6, 
-    "1000" : 0, 
-    "1020" : 1.6, 
-    "1040" : 4.0, 
-    "1050" : 4.8, 
-    "1070" : 6.2, 
-    "1090" : 7.3, 
-    "1110" : 8.3, 
-    "1130" : 9.4, 
-    "1150" : 10.5, 
-    "1170" : 11.6, 
-    "1190" : 12.6, 
-    "1210" : 13.6, 
-    "1230" : 14.7, 
-    "1250" : 15.8, 
-    "1270" : 16.9, 
-    "1290" : 18.1, 
-    "1310" : 19.1, 
-    "1330" : 20.6, 
-    "1350" : 21.6, 
-    "1370" : 22.9, 
-    "1390" : 24.1, 
-    "1410" : 25.1, 
-    "1430" : 26.6, 
-    "1450" : 27.6, 
-    "1460" : 29.6, 
-    "1480" : 30.6, 
-    "1500" : 31.6, 
-    "1520" : 33.6, 
-    "1540" : 34.6, 
-    "1560" : 36.6, 
-    "1580" : 37.6, 
-    "1600" : 39.6, 
-    "1620" : 40.6, 
-    "1640" : 42.6, 
-    "1660" : 43.6, 
-    "1680" : 45.6, 
-    "1700" : 47.6, 
-    "1720" : 48.6, 
-    "1740" : 50.6, 
-    "1760" : 51.6, 
-    "1780" : 53.6, 
-    "1800" : 55.6, 
-    "1820" : 57.6, 
-    "1840" : 59.6, 
-    "1860" : 61.6, 
-    "1880" : 62.6, 
-    "1890" : 65.6, 
-    "1910" : 67.6, 
-    "1930" : 69.6, 
-    "1950" : 71.6, 
-    "1970" : 73.6, 
-    "1990" : 75.6, 
-    "2010" : 78.6, 
-    "2030" : 80.6, 
-    "2050" : 83.6, 
-    "2070" : 85.6, 
-    "2090" : 88.6, 
-    "2110" : 90.6, 
-    "2130" : 93.6, 
-    "2150" : 96.6, 
-    "2170" : 98.6, 
-    "2190" : 101.6, 
-    "2210" : 104.6, 
-    "2230" : 107.6, 
-    "2250" : 109.6, 
-    "2270" : 112.6, 
-    "2290" : 117.6, 
-    "2300" : 118.6, 
-    "2320" : 121.6, 
-    "2340" : 124.6, 
-    "2360" : 127.6, 
-    "2380" : 130.6, 
-    "2400" : 134.6, 
-    "2420" : 137.6, 
-    "2440" : 140.6, 
-    "2460" : 144.6, 
-    "2480" : 147.6, 
-    "2500" : 151.6, 
-    "2520" : 154.6, 
-    "2540" : 158.6, 
-    "2560" : 162.6, 
-    "2580" : 166.6, 
-    "2600" : 170.6, 
-    "2620" : 174.6, 
-    "2640" : 179.6, 
-    "2660" : 183.6, 
-    "2680" : 187.6, 
-    "2700" : 192.6, 
-    "2710" : 196.6, 
-    "2730" : 201.6, 
-    "2750" : 206.6, 
-    "2770" : 210.6, 
-    "2790" : 215.6, 
-    "2810" : 220.6, 
-    "2830" : 225.6, 
-    "2850" : 230.6, 
-    "2870" : 236.6, 
-    "2890" : 241.6, 
-    "2910" : 246.6, 
-    "2930" : 252.6, 
-    "2950" : 257.6, 
-    "2970" : 263.6, 
-    "2990" : 268.6, 
-    "3010" : 274.6, 
-    "3030" : 280.6, 
-    "3050" : 286.6, 
-    "3070" : 292.6, 
-    "3090" : 299.6, 
-    "3110" : 305.6, 
-    "3130" : 311.6, 
-    "3140" : 318.6, 
-    "3160" : 324.6, 
-    "3180" : 331.6, 
-    "3200" : 338.6, 
-    "3220" : 345.6, 
-    "3240" : 352.6, 
-    "3260" : 359.6, 
-    "3280" : 367.6, 
-    "3300" : 374.6, 
-    "3320" : 382.6, 
-    "3340" : 389.6, 
-    "3360" : 397.6, 
-    "3380" : 405.6, 
-    "3400" : 413.6, 
-    "3420" : 421.6, 
-    "3440" : 430.6, 
-    "3460" : 438.6, 
-    "3480" : 447.6, 
-    "3500" : 455.6, 
-    "3520" : 464.6, 
-    "3540" : 472.5, 
-    "3550" : 483.6, 
-    "3570" : 492.6, 
-    "3590" : 502.6, 
-    "3610" : 512.6, 
-    "3630" : 522.6, 
-    "3650" : 532.6, 
-    "3670" : 542.6, 
-    "3690" : 552.6, 
-    "3710" : 563.6, 
-    "3730" : 574.6, 
-    "3750" : 585.6, 
-    "3770" : 596.6, 
-    "3790" : 607.6, 
-    "3810" : 619.6, 
-    "3830" : 631.6, 
-    "3850" : 643.6, 
-    "3870" : 655.6, 
-    "3890" : 667.6, 
-    "3910" : 679.6, 
-    "3930" : 692.6, 
-    "3950" : 704.6, 
-    "3960" : 718.5, 
-    "3980" : 731.6, 
-    "4000" : 744.6, 
-    "4020" : 758.6, 
-    "4040" : 772.6, 
-    "4060" : 786.6, 
-    "4080" : 801.6, 
-    "4100" : 815.6, 
-    "4120" : 830.6, 
-    "4140" : 846.6, 
-    "4160" : 861.6, 
-    "4180" : 876.6, 
-    "4200" : 892.6, 
-    "4220" : 908.6, 
-    "4240" : 924.6, 
-    "4260" : 941.6, 
-    "4280" : 958.6, 
-    "4300" : 975.6, 
-    "4320" : 993.6, 
-    "4340" : 1010.6, 
-    "4360" : 1029.6, 
-    "4380" : 1042.3, 
-    "4390" : 1065.6, 
-    "4410" : 1084.6, 
-    "4430" : 1103.6, 
-    "4450" : 1123.6, 
-    "4470" : 1143.6, 
-    "4490" : 1163.6, 
-    "4510" : 1184.6, 
-    "4530" : 1204.6, 
-    "4550" : 1226.6, 
-    "4570" : 1247.6, 
-    "4590" : 1269.6, 
-    "4610" : 1292.6, 
-    "4630" : 1314.6, 
-    "4650" : 1337.6, 
-    "4670" : 1360.6, 
-    "4690" : 1382.6, 
-    "4710" : 1408.6, 
-    "4730" : 1432.6, 
-    "4750" : 1457.6, 
-    "4770" : 1482.6, 
-    "4790" : 1508.6, 
-    "4800" : 1533.6, 
-    "4820" : 1559.6, 
-    "4840" : 1586.6, 
-    "4860" : 1613.6, 
-    "4880" : 1640.6, 
-    "4900" : 1669.6, 
-    "4920" : 1695.6, 
-    "4940" : 1723.6, 
-    "4960" : 1753.6, 
-    "4980" : 1805.5
-}
