--- ESP32/DIY-Flow-Bench/tools/mafDataDELPHI_AF10118.json
@@ -1,28 +1,0 @@
-    "0" : 0,
-    "600" : 5.620, 
-    "800" : 8.982, 
-    "940" : 10.711, 
-    "1000" : 12.440, 
-    "1300" : 19.597, 
-    "1400" : 22.382, 
-    "1500" : 25.456, 
-    "1600" : 29.251, 
-    "1700" : 33.526, 
-    "1800" : 38.137, 
-    "1900" : 43.036, 
-    "2000" : 48.175, 
-    "2100" : 53.987, 
-    "2200" : 59.222, 
-    "2400" : 74.112, 
-    "2600" : 91.259, 
-    "2800" : 111.144, 
-    "3000" : 132.950, 
-    "3200" : 158.262, 
-    "3400" : 187.177, 
-    "3600" : 217.724, 
-    "3800" : 258.503, 
-    "4000" : 295.631, 
-    "4250" : 342.461, 
-    "4500" : 421.040, 
-    "4750" : 490.925, 
-    "5000" : 490.925
