+++ platformio.ini
@@ -31,12 +31,10 @@ build_src_filter =
 	+<*.h> +<*.s> +<*.S> +<*.cpp> +<*.c> +<*.ino> +<src/> 
 	-<.git/> -<data/> -<test/> -<tests/> -<include/> -<mafData/> -<html/>
 extra_scripts = 
-	pre:user_actions_pre.py
-	post:user_actions_post.py
+	pre:ESP32/DIY-Flow-Bench/tools/user_actions_pre.py
+	post:ESP32/DIY-Flow-Bench/tools/user_actions_post.py
 
 	
-
-
 [common_env_data]
 lib_deps_builtin = 
 	DNSServer
@@ -72,6 +70,7 @@ build_flags =
  	-D CONFIG_ASYNC_TCP_RUNNING_CORE=1      ; force async_tcp task to be on same core as Arduino app (default is any core)
  	; -D CONFIG_ASYNC_TCP_STACK_SIZE=12288     ; reduce the stack size (default is 16K)
 
+
 ; General ESP32 build environment. Should work for most ESP32's
 [env:esp32dev]
 build_type = release
@@ -92,9 +91,6 @@ lib_deps =
 	; https://github.com/DeeEmm/BME680.git
 	majicdesigns/MD_REncoder@^1.0.1
 lib_ignore = 
-extra_scripts = 
-	pre:user_actions_pre.py
-	post:user_actions_post.py
 ; targets = clean, upload
